Remove argument checking when calling a function from MiniContext
Checking of call arguments was done based on a function's XML spec,
specifically, the `<params>` element - a form of dynamic type checking before each call.
There needed to be concordance between the spec
and the implementation (the spec wasn't just extra documentation).

This introduced friction when writing functions because, often, the implementation
would diverge from the spec. To make it easier for contributors to write functions
we`re moving towards dynamic argument checking/dispatching and XML specs/docs based
of language native (e.g. JsDoc for Js, roxygen for R, docstrings for Python).

In the future we will introduce more formal static type checking (based
on the entire call graph, not just single call arg checking). But for
now this change makes it easier to implement and use functions.